replaced api call and fixed camera

+++ app/src/main/java/com/example/wellfed/functionality/NutritionDataItem.kt
@@ -2,14 +2,14 @@ package com.example.wellfed.functionality
 
 data class NutritionDataItem(
     val name: String,
-    val calories: Int,
-    val serving_size_g: Int,
+    val calories: Double,
+    val serving_size_g: Double,
     val fat_total_g: Double,
     val fat_saturated_g: Double,
     val protein_g: Double,
-    val sodium_mg: Int,
-    val potassium_mg: Int,
-    val cholesterol_mg: Int,
+    val sodium_mg: Double,
+    val potassium_mg: Double,
+    val cholesterol_mg: Double,
     val carbohydrates_total_g: Double,
     val fiber_g: Double,
     val sugar_g: Double
